Store Go build cache and module cache in a volume
This change speeds up builds by storing the Go build/module caches in a
Docker volume. Previously these caches needed to be built up anew
every time that the container was created. We had a partial workaround
of downloading multi-life's module dependencies, which we no longer
need.

This change also clears the caches of data that is created while
installing vim-go.